feat: implement test-rule & refactor: change structure of the project
