+++ src/routes/graphql/index.ts
@@ -1,8 +1,9 @@
 import { FastifyPluginAsyncTypebox } from '@fastify/type-provider-typebox';
 import { createGqlResponseSchema, gqlResponseSchema } from './schemas.js';
 import { graphql, GraphQLSchema, parse, validate } from 'graphql';
-import { RootQuery } from './types/RootQuery.js';
-import { RootMutation } from './types/RootMutation.js';
+import depthLimit from 'graphql-depth-limit';
+import { RootQuery } from './schemas/rootQuery.js';
+import { RootMutation } from './schemas/rootMutation.js';
 
 const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
   fastify.route({
@@ -20,7 +21,7 @@ const plugin: FastifyPluginAsyncTypebox = async (fastify) => {
         mutation: RootMutation,
       });
 
-      const errors = validate(schema, parse(req.body.query));
+      const errors = validate(schema, parse(req.body.query), [depthLimit(5)]);
       if (errors.length > 0) return { data: null, errors: errors };
 
       return await graphql({
